Untabify KeyMapping files and rename/regroup classes one per file in a directory
