+++ TLM/TLM/TLM.csproj
@@ -176,7 +176,9 @@
     <Compile Include="State\ConfigData\TimedTrafficLights.cs" />
     <Compile Include="State\ConfigData\AdvancedVehicleAI.cs" />
     <Compile Include="State\GlobalConfig.cs" />
-    <Compile Include="State\KeyMapping.cs" />
+    <Compile Include="State\Keybinds\KeymappingSettings.cs" />
+    <Compile Include="State\Keybinds\KeymappingSettingsMain.cs" />
+    <Compile Include="State\Keybinds\TmpeRebindableKey.cs" />
     <Compile Include="TrafficLight\FlowWaitCalcMode.cs" />
     <Compile Include="TrafficLight\ICustomSegmentLight.cs" />
     <Compile Include="TrafficLight\Impl\CustomSegmentLights.cs" />
