+++ TLM/TLM/UI/MainMenu/PrioritySignsButton.cs
@@ -1,5 +1,6 @@
 ﻿using ColossalFramework;
 using TrafficManager.State;
+using TrafficManager.State.Keybinds;
 
 namespace TrafficManager.UI.MainMenu {
     public class PrioritySignsButton : MenuToolModeButton {
@@ -7,6 +8,6 @@ namespace TrafficManager.UI.MainMenu {
         public override ButtonFunction Function => ButtonFunction.PrioritySigns;
         public override string Tooltip => "Add_priority_signs";
         public override bool Visible => Options.prioritySignsEnabled;
-        public override SavedInputKey ShortcutKey => OptionsKeymapping.KeyPrioritySignsTool;
+        public override SavedInputKey ShortcutKey => KeymappingSettings.KeyPrioritySignsTool;
     }
 }
