+++ TLM/TLM/UI/MainMenu/SpeedLimitsButton.cs
@@ -1,5 +1,6 @@
 ﻿using ColossalFramework;
 using TrafficManager.State;
+using TrafficManager.State.Keybinds;
 
 namespace TrafficManager.UI.MainMenu {
     public class SpeedLimitsButton : MenuToolModeButton {
@@ -7,6 +8,6 @@ namespace TrafficManager.UI.MainMenu {
         public override ButtonFunction Function => ButtonFunction.SpeedLimits;
         public override string Tooltip => "Speed_limits";
         public override bool Visible => Options.customSpeedLimitsEnabled;
-        public override SavedInputKey ShortcutKey => OptionsKeymapping.KeySpeedLimitsTool;
+        public override SavedInputKey ShortcutKey => KeymappingSettings.KeySpeedLimitsTool;
     }
 }
