+++ TLM/TLM/UI/MainMenu/VehicleRestrictionsButton.cs
@@ -5,6 +5,7 @@ namespace TrafficManager.UI.MainMenu {
         public override ToolMode ToolMode => ToolMode.VehicleRestrictions;
         public override ButtonFunction Function => ButtonFunction.VehicleRestrictions;
         public override string Tooltip => "Vehicle_restrictions";
+
         public override bool Visible => Options.vehicleRestrictionsEnabled;
     }
 }
